requirements.txt
pystray
Pillow
pythonnet